Added src directory and moved app and logic to src
